requirements.txt
blinker==1.6.2
click==8.1.3
dominate==2.7.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
visitor==0.1.3
Werkzeug==2.3.3
